Add initial design doc
It's important to keep track of decisions around the project and what we
want our goals to be. This prevents feature creep and makes sure that
we're focussing on the desired behaviour as a goal. The design doc is
where we outline this.

I was also getting some annoying markdownlint warnings that kept being
triggered by some default formatting VS Code wanted to implement. Rather
than trying to fix that, I've just turned off the rules that were
causing problems. I don't think for this project we need towrry about
line length in markdown or how many spaces we enforce for lists. If
people feel otherwise, they're welcome to edit this as needed.